backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1.1
gunicorn==20.0.4
convertdate==2.1.3
python-dateutil==2.8.1
psycopg2-binary==2.8.6
ptvsd==4.3.2
pymysql==0.10.1
django-mysql==2.2.0
djangorestframework==3.12.2
django-cors-headers==3.3.0
requests==2.25.0
unittest-xml-reporting==3.0.4
requests==2.25.0
cryptography==2.8
drf-jwt==1.13.3
pyjwt==1.7.1
